added stm32 uart generic mechanism and uart3 device driver. fix problem with interrupt priority (see embos documentation) priority level must be > 128. same pb on lm3S. added embOS cortex M3/M4 version 3.86n.

+++ sys/root/src/kernel/core/types.h
@@ -104,7 +104,9 @@ typedef int __off_t;
    #pragma message ("note: use lepton time_t definition")
 #endif
    #if (__KERNEL_COMPILER_SUPPORT_TYPE>__KERNEL_COMPILER_SUPPORT_32_BITS_TYPE)
-      typedef uint64_t __time_t;
+       //to do time64 bits
+      //typedef uint64_t __time_t;
+      typedef uint32_t __time_t;
    #else
       typedef uint32_t __time_t;
    #endif
